tests overhaul
