--- fromscratchtoml/test/__init__.py
@@ -1,1 +1,0 @@
-from .neural_network.grad_check import Gradient_check # noqa:F401
