--- fromscratchtoml/test/neural_network/__init__.py
@@ -1,1 +1,0 @@
-from .grad_check import Gradient_check # noqa:F401
